+++ client/lib/checkmate-scan.js
@@ -1,0 +1,28 @@
+import Coords from './coords';
+import isViableStart from './is-viable-start';
+
+const coords = new Coords();
+
+export default function checkmateScan(board, gamestate) {
+  const enemyCoords = [];
+
+  // find location of all enemies
+  for (const coord of coords) {
+    if (board[coord]) {
+      if (board[coord].player === gamestate.turn[1]) {
+        enemyCoords.push(coord);
+      }
+    }
+  }
+
+  // return if there is no checkmate
+  for (const enemyCoord of enemyCoords) {
+    if (isViableStart(board, gamestate, enemyCoord, gamestate.nextTurn)) {
+      return;
+    }
+  }
+
+  // otherwise checkmate
+  gamestate.checkmate = true;
+  console.log('Checkmate!!'); // eslint-disable-line
+}
