+++ client/lib/copy.js
@@ -1,0 +1,3 @@
+export default function copy(obj) {
+  return { ...obj };
+}
