+++ client/lib/empty.js
@@ -1,0 +1,6 @@
+export default function empty(board, coord) {
+  board[coord] = {
+    piece: null,
+    player: null
+  };
+}
