+++ client/lib/is-empty-at.js
@@ -1,0 +1,7 @@
+export default function isEmptyAt(board, coord) {
+  if (!board[coord].piece) {
+    return true;
+  } else {
+    return false;
+  }
+}
