+++ client/lib/pawn-scan.js
@@ -1,0 +1,15 @@
+export default function pawnScan(board, gamestate) {
+  for (let i = 81; i < 89; i++) {
+    if (board[i].piece === 'p') {
+      gamestate.promoting = board[i];
+      return;
+    }
+  }
+
+  for (let i = 11; i < 19; i++) {
+    if (board[i].piece === 'p') {
+      gamestate.promoting = board[i];
+      return;
+    }
+  }
+}
